Move notification to db package and create a DAO
