+++ app/src/main/java/org/wikipedia/analytics/NotificationInteractionFunnel.kt
@@ -5,7 +5,7 @@ import androidx.core.app.NotificationManagerCompat
 import org.json.JSONObject
 import org.wikipedia.Constants
 import org.wikipedia.WikipediaApp
-import org.wikipedia.notifications.Notification
+import org.wikipedia.notifications.db.Notification
 import org.wikipedia.notifications.NotificationPollBroadcastReceiver
 
 class NotificationInteractionFunnel(app: WikipediaApp, private val id: Long, private val wiki: String, private val type: String?) : Funnel(app, SCHEMA_NAME, REV_ID) {
