+++ app/src/main/java/org/wikipedia/views/NotificationActionsOverflowView.kt
@@ -23,7 +23,7 @@ import org.wikipedia.analytics.NotificationInteractionFunnel
 import org.wikipedia.analytics.eventplatform.NotificationInteractionEvent
 import org.wikipedia.databinding.ViewNotificationActionsOverflowBinding
 import org.wikipedia.dataclient.WikiSite
-import org.wikipedia.notifications.Notification
+import org.wikipedia.notifications.db.Notification
 import org.wikipedia.notifications.NotificationCategory
 import org.wikipedia.notifications.NotificationLinkHandler
 import org.wikipedia.notifications.NotificationListItemContainer
